requirements.txt
astroid==2.3.3
attrs==19.3.0
Authlib==0.13
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-JSONRPC==0.3.1
Flask-SocketIO==4.2.1
gevent==1.3.7
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycparser==2.19
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.2
pytest-flask==0.15.0
python-engineio==3.9.3
python-socketio==4.3.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
